feat!: migrate challenges to using state conditions
Task Goals do NOT work how you think they work. It's very annoying. The only way to properly evaluate for multiple conditions is to use the State Conditions.

I do not know why there are two ways to define goals. It doesn't make sense to me, but i only care that this works