refactor: simplify how trajectory are run per use-case
This is a lot of lines for the refactor, but this should make it easier to separate trajectory generation from evaluation in an easy way.

The WandB trackers have been massively overhauled so that they run with different projects.

This also brings in all the various "patches" that were needed to run the last round of eval ai